fix: add separate cjs and esm bundles
