+++ src/ambient.d.ts
